requirements.txt
numpy
scipy
opencv-python
joblib
Pillow